Revert "Revert "Revert "Merge branch 'master' of https://github.com/Eytins/HotelManagement"""
This reverts commit 442a2f0afbec22d9aefc52283811dbe453fc06f7.

+++ src/com/HotelManagement/dao/hotelType/HotelTypeMapper.java
@@ -33,7 +33,7 @@ public interface HotelTypeMapper {
 
     //更改房型
     int modifyRoomType(@Param("roomTypeId") Integer roomTypeId,
-                       @Param("price") long price,
+                       @Param("price") Integer price,
                        @Param("numberOfPerson") Integer numberOfPerson);
